@@ -346,12 +346,7 @@ public final class ZipTestCase extends AbstractTestCase {
     }
     String first_payload = "ABBA";
     String second_payload = "AAAAAAAAAAAA";
-    ZipArchiveEntryPredicate allFilesPredicate = new ZipArchiveEntryPredicate() {
-        @Override
-        public boolean test(final ZipArchiveEntry zipArchiveEntry) {
-            return true;
-        }
-    };
+    ZipArchiveEntryPredicate allFilesPredicate = zipArchiveEntry -> true;
 
     @Test
     public void testCopyRawEntriesFromFile()
